refactor: move whole section components to `sections` directory
