+++ src/pages/Home.jsx
@@ -1,8 +1,8 @@
 import Header from '../components/Header.jsx';
-import Hero from '../components/Hero.jsx';
-import OurStory from '../components/OurStory.jsx';
-import HowWeWork from '../components/HowWeWork.jsx';
-import Projects from '../components/Projects.jsx';
+import Hero from '../components/sections/Hero.jsx';
+import OurStory from '../components/sections/OurStory.jsx';
+import HowWeWork from '../components/sections/HowWeWork.jsx';
+import Projects from '../components/sections/Projects.jsx';
 
 function Home() {
   return (
